requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
boto3==1.36.17
botocore==1.36.17
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
Django==5.1.5
django-cors-headers==4.6.0
django-environ==0.12.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
docopt==0.6.2
drf-yasg==1.21.8
fonttools==4.56.0
frozenlist==1.5.0
git-filter-repo==2.47.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langsmith==0.3.8
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
mongoengine==0.29.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.2
orjson==3.10.15
packaging==24.2
pillow==11.1.0
pipreqs==0.4.13
propcache==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
tenacity==9.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
yarg==0.1.10
yarl==1.18.3
zstandard==0.23.0
seaborn==0.13.2